requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary